http: remove k8s deps + extraneous features
Biggest removal here is ability to follow non-local redirects. As of
Kubernetes v1.20.2, this is _always_ off via a compile-time
constant [1], so it didn't make sense to support.

The biggest limitation is that some subtlety around HTTP/2
configuration has been lost and should be revisited but is not
critical for an initial version.

Most of the (extremely verbose) log statements have been kept
but commented out; a decision should be made on how to handle
these (they can likely be safely dropped entirely for our use
cases).

[1]: https://github.com/kubernetes/kubernetes/blob/v1.20.2/pkg/kubelet/prober/prober.go#L68-L73