exec: remove k8s deps + update import path
This is the roughest of the three probes and will evolve/change
substantially -- this is simply the minimal viable to move it away
from exec'ing in a container to native OS.

Special case error-handling for an exec timeout in the container
is gone; in the future, usage of `exec.CommandContext` to create
the command coupled with `context.WithTimeout` can ensure this.
(It's actually a general problem across all three probe
implementations currently).

Additionally, the K8s version limited the amount of output captured
from stdout/stderr to 10KB; no such limit exists anymore due to the
need to port `ioutils.LimitedWriter` from K8s. This can be done as
a future improvement because it's not critical for an initial version.