worker: use ProbeFunc instead of interface
This will be paired with a prober manager that can keep a single
instance of each prober to be invoked for each probe by wrapping
in a func to meet the `Probe` signature. This is a good middle
ground between having `Worker` know TOO much and forcing a 1:1
lifecycle of prober:worker.

It's simpler than K8s model because over there the worker is
actually passed the prober manager and then calls back into it with
the relevant container + options and the prober manager does a LOT.
(This approach doesn't prevent from doing more, the functions can
always be wrapped as necessary.)