probe: use context instead of explicit timeout args (#10)
The underlying TCP/HTTP code in Go stdlib uses context to apply the
timeouts anyway, so this doesn't affect things much. The biggest
motivation is that `exec.CommandContext` is far and away the most
straightforward way to limit how long a process can execute for and
have it auto-killed, so while we could take a timeout arg there and
make a context with it, it seems reasonable to make everything
consistent and give the caller full control over probe lifetime.