prober: prober manager to create ProberFunc from shared probers (#13)
This is an extremely basic analogue to the probe manager from K8s
in that it initializes the known prober types (HTTP/TCP/exec) so
that all probes can share underlying instances.

In K8s, however, the prober manager does a lot of other things (such
as sending container events) and is actually invoked _by_ the worker.
Here, the prober manager just emits `ProberFunc` wrappers that use
the shared prober instances. This keeps the worker more generic and
decoupled but provides the same lifecycle benefits.