exec: kill process group to avoid dangling processes (#18)
This is now fairly similar to how we actually run processes in
Tilt.

The high-level approach is:
 * Use `Cmd::Start()` + `Cmd::Wait()`
 * Monitor `Cmd::Wait()` in goroutine and write to channel on done
 * Block on `Context::Done()` and cmd wait channel
 * Kill process group and return error if context done first
 * Return normally if `Cmd` finishes normally

The diff has a ton of deleted code because the K8s exec wrapper
doesn't expose `SysprocAttr`, so it wasn't possible to use with
process groups. A simpler interface is now provided, and there's
an integration test for the behavior there.

There's also a new test to verify that overlapping probes are not
run, which originally seemed to be the cause of lots of dangling
processes. The worker code is actually correct, so this serves as
a reasonable regression test to ensure that doesn't happen in the
future.